Refactor classes in the `testutil` module.
