+++ api-java/src/test/java/org/spine3/examples/todolist/c/aggregate/TaskLabelsPartTest.java
@@ -42,6 +42,7 @@ import org.spine3.examples.todolist.c.failures.CannotAssignLabelToTask;
 import org.spine3.examples.todolist.c.failures.CannotRemoveLabelFromTask;
 import org.spine3.examples.todolist.context.TodoListBoundedContext;
 import org.spine3.examples.todolist.testdata.TestResponseObserver;
+import org.spine3.examples.todolist.testdata.TestTaskLabelsCommandFactory;
 import org.spine3.server.BoundedContext;
 import org.spine3.server.command.CommandBus;
 
@@ -53,11 +54,11 @@ import static org.spine3.base.Commands.create;
 import static org.spine3.base.Identifiers.newUuid;
 import static org.spine3.examples.todolist.testdata.TestCommandContextFactory.createCommandContext;
 import static org.spine3.examples.todolist.testdata.TestTaskCommandFactory.DESCRIPTION;
-import static org.spine3.examples.todolist.testdata.TestTaskCommandFactory.assignLabelToTaskInstance;
+import static org.spine3.examples.todolist.testdata.TestTaskLabelsCommandFactory.assignLabelToTaskInstance;
 import static org.spine3.examples.todolist.testdata.TestTaskCommandFactory.completeTaskInstance;
 import static org.spine3.examples.todolist.testdata.TestTaskCommandFactory.createTaskInstance;
 import static org.spine3.examples.todolist.testdata.TestTaskCommandFactory.deleteTaskInstance;
-import static org.spine3.examples.todolist.testdata.TestTaskCommandFactory.removeLabelFromTaskInstance;
+import static org.spine3.examples.todolist.testdata.TestTaskLabelsCommandFactory.removeLabelFromTaskInstance;
 
 /**
  * @author Illia Shepilov
@@ -108,7 +109,7 @@ public class TaskLabelsPartTest {
         @Test
         @DisplayName("produce LabelAssignedToTask event")
         public void produceEvent() {
-            final AssignLabelToTask assignLabelToTaskCmd = assignLabelToTaskInstance(taskId, labelId);
+            final AssignLabelToTask assignLabelToTaskCmd = TestTaskLabelsCommandFactory.assignLabelToTaskInstance(taskId, labelId);
             final List<? extends Message> messageList =
                     taskLabelsPart.dispatchForTest(assignLabelToTaskCmd, COMMAND_CONTEXT);
 
@@ -133,7 +134,7 @@ public class TaskLabelsPartTest {
                 final Command deleteTaskCmd = create(deleteTask, COMMAND_CONTEXT);
                 commandBus.post(deleteTaskCmd, responseObserver);
 
-                final AssignLabelToTask assignLabelToTask = assignLabelToTaskInstance(taskId, labelId);
+                final AssignLabelToTask assignLabelToTask = TestTaskLabelsCommandFactory.assignLabelToTaskInstance(taskId, labelId);
                 final Command assignLabelToTaskCmd = create(assignLabelToTask, COMMAND_CONTEXT);
                 commandBus.post(assignLabelToTaskCmd, responseObserver);
             } catch (Throwable e) {
@@ -155,7 +156,7 @@ public class TaskLabelsPartTest {
                 final Command completeTaskCmd = create(completeTask, COMMAND_CONTEXT);
                 commandBus.post(completeTaskCmd, responseObserver);
 
-                final AssignLabelToTask assignLabelToTask = assignLabelToTaskInstance(taskId, labelId);
+                final AssignLabelToTask assignLabelToTask = TestTaskLabelsCommandFactory.assignLabelToTaskInstance(taskId, labelId);
                 final Command assignLabelToTaskCmd = Commands.create(assignLabelToTask, COMMAND_CONTEXT);
                 commandBus.post(assignLabelToTaskCmd, responseObserver);
             } catch (Throwable e) {
@@ -173,7 +174,7 @@ public class TaskLabelsPartTest {
         @Test
         @DisplayName("produce LabelRemovedFromTask event")
         public void produceEvent() {
-            final RemoveLabelFromTask removeLabelFromTaskCmd = removeLabelFromTaskInstance(taskId, labelId);
+            final RemoveLabelFromTask removeLabelFromTaskCmd = TestTaskLabelsCommandFactory.removeLabelFromTaskInstance(taskId, labelId);
             final List<? extends Message> messageList =
                     taskLabelsPart.dispatchForTest(removeLabelFromTaskCmd, COMMAND_CONTEXT);
 
@@ -198,7 +199,7 @@ public class TaskLabelsPartTest {
                 final Command completeTaskCmd = create(completeTask, COMMAND_CONTEXT);
                 commandBus.post(completeTaskCmd, responseObserver);
 
-                final RemoveLabelFromTask removeLabelFromTask = removeLabelFromTaskInstance(taskId, labelId);
+                final RemoveLabelFromTask removeLabelFromTask = TestTaskLabelsCommandFactory.removeLabelFromTaskInstance(taskId, labelId);
                 final Command removeLabelFromTaskCmd = create(removeLabelFromTask, COMMAND_CONTEXT);
                 commandBus.post(removeLabelFromTaskCmd, responseObserver);
             } catch (Throwable e) {
@@ -220,7 +221,7 @@ public class TaskLabelsPartTest {
                 final Command deleteTaskCmd = create(deleteTask, COMMAND_CONTEXT);
                 commandBus.post(deleteTaskCmd, responseObserver);
 
-                final RemoveLabelFromTask removeLabelFromTask = removeLabelFromTaskInstance(taskId, labelId);
+                final RemoveLabelFromTask removeLabelFromTask = TestTaskLabelsCommandFactory.removeLabelFromTaskInstance(taskId, labelId);
                 final Command removeLabelFromTaskCmd = create(removeLabelFromTask, COMMAND_CONTEXT);
                 commandBus.post(removeLabelFromTaskCmd, responseObserver);
             } catch (Throwable e) {
