+++ client/src/test/java/org/spine3/examples/todolist/client/CommandLineTodoClientTest.java
@@ -43,7 +43,7 @@ import java.util.concurrent.TimeUnit;
 import static org.spine3.base.Identifiers.newUuid;
 import static org.spine3.client.ConnectionConstants.DEFAULT_CLIENT_SERVICE_PORT;
 import static org.spine3.examples.todolist.testdata.TestTaskCommandFactory.DESCRIPTION;
-import static org.spine3.examples.todolist.testdata.TestTaskLabelCommandFactory.LABEL_TITLE;
+import static org.spine3.examples.todolist.testdata.TestLabelCommandFactory.LABEL_TITLE;
 
 /**
  * @author Illia Shepilov
