+++ client/src/test/java/org/spine3/examples/todolist/client/UpdateTaskPriorityTest.java
@@ -33,12 +33,13 @@ import org.spine3.examples.todolist.c.commands.CreateDraft;
 import org.spine3.examples.todolist.c.commands.UpdateTaskPriority;
 import org.spine3.examples.todolist.q.projection.LabelledTasksView;
 import org.spine3.examples.todolist.q.projection.TaskView;
+import org.spine3.examples.todolist.testdata.TestTaskLabelsCommandFactory;
 
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static org.spine3.examples.todolist.testdata.TestTaskCommandFactory.assignLabelToTaskInstance;
+import static org.spine3.examples.todolist.testdata.TestTaskLabelsCommandFactory.assignLabelToTaskInstance;
 import static org.spine3.examples.todolist.testdata.TestTaskCommandFactory.updateTaskPriorityInstance;
 
 /**
@@ -148,7 +149,7 @@ public class UpdateTaskPriorityTest extends CommandLineTodoClientTest {
         client.create(createLabel);
 
         final int expectedListSize = 1;
-        final AssignLabelToTask assignLabelToTask = assignLabelToTaskInstance(createdTaskId, createLabel.getLabelId());
+        final AssignLabelToTask assignLabelToTask = TestTaskLabelsCommandFactory.assignLabelToTaskInstance(createdTaskId, createLabel.getLabelId());
         client.assignLabel(assignLabelToTask);
 
         updatePriority(priority, isCorrectId, createdTaskId);
