+++ testutil/src/main/java/org/spine3/examples/todolist/testdata/TestEventEnricherFactory.java
@@ -33,10 +33,10 @@ import org.spine3.server.event.enrich.EventEnricher;
 
 import java.util.function.Function;
 
-import static org.spine3.examples.todolist.testdata.TestEventFactory.LABEL_ID;
+import static org.spine3.examples.todolist.testdata.TestTaskEventFactory.LABEL_ID;
 
 /**
- * Provides event enricher for test needs.
+ * Provides event enricher for the test needs.
  *
  * @author Illia Shepilov
  */
